Revert "web: Avoid the use of float arguments."
This work-around was specific to Emterpreter, which is obsoleted by ASYNCIFY.

This reverts commit dce59141a7f940da9f5b2820ee4b61016b9113b3.

+++ renpy/gl/glenviron_shader.pyx
@@ -541,7 +541,7 @@ cdef class ShaderEnviron(Environ):
         else:
             glDisableVertexAttribArray(tex)
 
-    cdef void set_color(self, double r, double g, double b, double a):
+    cdef void set_color(self, float r, float g, float b, float a):
         glUniform4f(self.program.Color, r, g, b, a)
 
     cdef void ortho(self, double left, double right, double bottom, double top, double near, double far):
